Move Options::Include to separate source file
* Extract method refactoring to break up long `.call`.
